apps/srvc-inference/tf_idf/requirements.txt
kafka-python==2.0.2
qdrant-client==1.7.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
mlflow==2.9.1
